refactor(qbittorrent-e2e): extract client role module
