+++ src/console/ci/qbittorrent/mod.rs
@@ -1,4 +1,5 @@
 pub mod bencode;
+pub mod client_role;
 pub mod qbittorrent_client;
 pub mod runner;
 pub mod torrent_artifacts;
